DEV: When click on categories go directly to their edit page (#3)
Chat still relies on categories for permissions, but we don't want to
enter categories where it shows a list of topics. Instead when a
category is clicked on go directly to the category edit page.

This way we get direct access through chat when editing a channel to
also edit the category permissions tied to that channel.